added radars to player and AI struct, made a new class called , made a seperate class to manage ai called AIManager
