+++ VTOLVR-Multiplayer/VTOLVR-Multiplayer.csproj
@@ -94,6 +94,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AIManager.cs" />
     <Compile Include="HPInfo.cs" />
     <Compile Include="Messages\Message.cs" />
     <Compile Include="Messages\MessageType.cs" />
